+++ src/v5_hal/firmware/src/NodeManager.cpp
@@ -1,12 +1,13 @@
 #include "NodeManager.h"
 
-NodeManager::NodeManager(long unsigned int (*getMilliseconds)(void)) { 
-    m_getMillis = getMilliseconds; 
+NodeManager::NodeManager(uint32_t (*getMilliseconds)(void)) {
+    m_getMillis = getMilliseconds;
     m_handle = new ros::NodeHandle();
 }
 
-ros::NodeHandle* NodeManager::addNode(Node* node, unsigned long intervalMilliseconds) {
-    NodeManager::NodeStructure nodeStructure = {node, intervalMilliseconds, 0UL};
+ros::NodeHandle* NodeManager::addNode(Node* node,
+                                      uint32_t intervalMilliseconds) {
+    NodeManager::NodeStructure nodeStructure = {node, intervalMilliseconds, 0};
     m_nodeStructures.push_back(nodeStructure);
 
     return m_handle;
@@ -19,7 +20,7 @@ void NodeManager::initialize() {
 }
 
 void NodeManager::execute() {
-    for (auto &nodeStructure : m_nodeStructures) {
+    for (auto& nodeStructure : m_nodeStructures) {
         auto currentTime = m_getMillis();
         if (currentTime - nodeStructure.lastExecutedMillis >=
             nodeStructure.triggerMillis) {
@@ -30,6 +31,4 @@ void NodeManager::execute() {
     pros::c::delay(m_delayTimeMillis);
 }
 
-NodeManager::~NodeManager() { 
-    m_nodeStructures.clear(); 
-}
+NodeManager::~NodeManager() { m_nodeStructures.clear(); }
